pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.12,<3.13"
name = "ml_bike_demand_end_to_end"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "ipython>=8.10",
    "jupyterlab>=3.0",
    "notebook",
    "kedro~=1.1",
    "kedro-datasets>=1.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "catboost>=1.2",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
    "dash>=2.14",
    "dash-bootstrap-components>=1.5",
    "plotly>=5.18",
    "matplotlib>=3.8",
    "kedro-viz>=12.3.0",
    "kedro-mlflow>=2.0.2",
    "fastapi>=0.136.0",
    "uvicorn>=0.45.0",
    "requests>=2.33.1",
]

[project.scripts]
"ml-bike-demand-end-to-end" = "ml_bike_demand_end_to_end.__main__:main"

[project.entry-points."kedro.hooks"]

[project.optional-dependencies]
docs = [
    "docutils<0.21",
    "sphinx>=5.3,<7.3",
    "sphinx_rtd_theme==2.0.0",
    "nbsphinx==0.8.1",
    "sphinx-autodoc-typehints==1.20.2",
    "sphinx_copybutton==0.5.2",
    "ipykernel>=5.3, <7.0",
    "Jinja2<3.2.0",
    "myst-parser>=1.0,<2.1"
]
dev = [
    "pytest-cov>=3,<7",
    "pytest-mock>=1.7.1, <2.0",
    "pytest~=7.2",
    "ruff~=0.12.0"
]

[tool.setuptools.dynamic]
version = {attr = "ml_bike_demand_end_to_end.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.kedro]
package_name = "ml_bike_demand_end_to_end"
project_name = "ml-bike-demand-end-to-end"
kedro_init_version = "1.3.1"
tools = "['Linting', 'Testing', 'Custom Logging', 'Documentation', 'Data Structure']"
example_pipeline = "False"
source_dir = "src"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/ml_bike_demand_end_to_end -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.ruff.format]
docstring-code-format = true

[tool.ruff]
line-length = 88
show-fixes = true

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "W",   # pycodestyle
    "E",   # pycodestyle
    "I",   # isort
    "UP",  # pyupgrade
    "PL",  # Pylint
    "T201", # Print Statement
]
ignore = ["E501"]  # Ruff format takes care of line-too-long

[tool.kedro_telemetry]
project_id = "4a21c5ec379d4d23a1dbaa02561f39a7"
